+++ packages/runtime/src/helpers/dev.ts
@@ -1,8 +1,11 @@
+import type { Buffer } from 'buffer'
 import { resolve } from 'path'
+import { Transform } from 'stream'
 
 import { OnPreBuild } from '@netlify/build'
 import execa from 'execa'
-import { unlink, existsSync } from 'fs-extra'
+import { unlink } from 'fs-extra'
+import mergeStream from 'merge-stream'
 
 import { writeDevEdgeFunction } from './edge'
 import { patchNextFiles } from './files'
@@ -20,21 +23,31 @@ export const onPreDev: OnPreBuild = async ({ constants, netlifyConfig }) => {
   })
   await writeDevEdgeFunction(constants)
 
-  if (!existsSync(resolve(base, 'middleware.ts')) && !existsSync(resolve(base, 'middleware.js'))) {
-    console.log(
-      "No middleware found. If you did intend to use middleware, create a 'middleware.ts' or 'middleware.js' file in your project root to add custom middleware.",
-    )
-  } else {
-    console.log('Watching for changes in Next.js middleware...')
-
-    // Eventually we might want to do this via esbuild's API, but for now the CLI works fine
-    const common = [`--bundle`, `--outdir=${resolve('.netlify')}`, `--format=esm`, `--target=esnext`, '--watch']
-
-    // TypeScript
-    execa(`esbuild`, [...common, resolve(base, 'middleware.ts')], { all: true }).all.pipe(process.stdout)
-
-    // JavaScript
-    execa(`esbuild`, [...common, resolve(base, 'middleware.js')], { all: true }).all.pipe(process.stdout)
-    // Don't return the promise because we don't want to wait for the child process to finish
+  // Eventually we might want to do this via esbuild's API, but for now the CLI works fine
+  const common = [`--bundle`, `--outdir=${resolve('.netlify')}`, `--format=esm`, `--target=esnext`, '--watch']
+  const opts = {
+    all: true,
+    env: { ...process.env, FORCE_COLOR: '1' },
   }
+  // TypeScript
+  const tsout = execa(`esbuild`, [...common, resolve(base, 'middleware.ts')], opts).all
+
+  // JavaScript
+  const jsout = execa(`esbuild`, [...common, resolve(base, 'middleware.js')], opts).all
+
+  const filter = new Transform({
+    transform(chunk: Buffer, encoding, callback) {
+      const str = chunk.toString(encoding)
+
+      // Skip if message includes this, because we run even when the files are missing
+      if (!str.includes('[ERROR] Could not resolve')) {
+        this.push(chunk)
+      }
+      callback()
+    },
+  })
+
+  mergeStream(tsout, jsout).pipe(filter).pipe(process.stdout)
+
+  // Don't return the promise because we don't want to wait for the child process to finish
 }